style(global): separating proxy from dledger
1. separating proxy from dledger
