+++ command/src/main/java/io/openmessaging/storage/dledger/DLedger.java
@@ -19,9 +19,9 @@ package io.openmessaging.storage.dledger;
 import com.alibaba.fastjson.JSON;
 import com.beust.jcommander.JCommander;
 import io.openmessaging.storage.dledger.cmdline.ConfigCommand;
-import io.openmessaging.storage.dledger.dledger.DLedgerProxy;
-import io.openmessaging.storage.dledger.dledger.DLedgerProxyConfig;
-import io.openmessaging.storage.dledger.utils.ConfigUtils;
+import io.openmessaging.storage.dledger.proxy.DLedgerProxy;
+import io.openmessaging.storage.dledger.proxy.DLedgerProxyConfig;
+import io.openmessaging.storage.dledger.proxy.util.ConfigUtils;
 import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
