+++ command/src/main/java/io/openmessaging/storage/dledger/cmdline/ServerCommand.java
@@ -20,8 +20,8 @@ import com.beust.jcommander.Parameter;
 import com.beust.jcommander.Parameters;
 import io.openmessaging.storage.dledger.DLedger;
 import io.openmessaging.storage.dledger.DLedgerConfig;
-import io.openmessaging.storage.dledger.dledger.DLedgerProxyConfig;
-import io.openmessaging.storage.dledger.utils.ConfigUtils;
+import io.openmessaging.storage.dledger.proxy.DLedgerProxyConfig;
+import io.openmessaging.storage.dledger.proxy.util.ConfigUtils;
 import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
