+++ command/src/test/resources/config.example.error.yaml
