+++ command/src/test/resources/config.example.yaml
