+++ dledger/src/test/java/io/openmessaging/storage/dledger/BatchPushTest.java
@@ -17,7 +17,6 @@
 package io.openmessaging.storage.dledger;
 
 import io.openmessaging.storage.dledger.client.DLedgerClient;
-import io.openmessaging.storage.dledger.dledger.DLedgerProxy;
 import io.openmessaging.storage.dledger.entry.DLedgerEntry;
 import io.openmessaging.storage.dledger.protocol.AppendEntryRequest;
 import io.openmessaging.storage.dledger.protocol.AppendEntryResponse;
@@ -38,14 +37,13 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doAnswer;
 
 public class BatchPushTest extends ServerTestHarness{
-
     @Test
     public void testBatchPushWithOneByOneRequests() throws Exception {
         String group = UUID.randomUUID().toString();
         String peers = String.format("n0-localhost:%d;n1-localhost:%d;n2-localhost:%d", nextPort(), nextPort(), nextPort());
-        DLedgerProxy dLedgerProxy0 = launchDLedgerProxyEnableBatchPush(group, peers, "n0", "n1", DLedgerConfig.FILE);
-        DLedgerProxy dLedgerProxy1 = launchDLedgerProxyEnableBatchPush(group, peers, "n1", "n1", DLedgerConfig.FILE);
-        DLedgerProxy dLedgerProxy2 = launchDLedgerProxyEnableBatchPush(group, peers, "n2", "n1", DLedgerConfig.FILE);
+        DLedgerServer dLedgerServer0 = launchServerEnableBatchPush(group, peers, "n0", "n1", DLedgerConfig.FILE);
+        DLedgerServer dLedgerServer1 = launchServerEnableBatchPush(group, peers, "n1", "n1", DLedgerConfig.FILE);
+        DLedgerServer dLedgerServer2 = launchServerEnableBatchPush(group, peers, "n2", "n1", DLedgerConfig.FILE);
         DLedgerClient dLedgerClient = launchClient(group, peers);
         for (int i = 0; i < 10; i++) {
             AppendEntryResponse appendEntryResponse = dLedgerClient.append(("testBulkCopyWithOneByOneRequests" + i).getBytes());
@@ -53,13 +51,9 @@ public class BatchPushTest extends ServerTestHarness{
             Assertions.assertEquals(i, appendEntryResponse.getIndex());
         }
         Thread.sleep(100);
-
-        DLedgerServer dLedgerServer0 = dLedgerProxy0.getDLedgerManager().getDLedgerServers().get(0);
-        DLedgerServer dLedgerServer1 = dLedgerProxy1.getDLedgerManager().getDLedgerServers().get(0);
-        DLedgerServer dLedgerServer2 = dLedgerProxy2.getDLedgerManager().getDLedgerServers().get(0);
-        Assertions.assertEquals(9, dLedgerServer0.getdLedgerStore().getLedgerEndIndex());
-        Assertions.assertEquals(9, dLedgerServer1.getdLedgerStore().getLedgerEndIndex());
-        Assertions.assertEquals(9, dLedgerServer2.getdLedgerStore().getLedgerEndIndex());
+        Assertions.assertEquals(9, dLedgerServer0.getDLedgerStore().getLedgerEndIndex());
+        Assertions.assertEquals(9, dLedgerServer1.getDLedgerStore().getLedgerEndIndex());
+        Assertions.assertEquals(9, dLedgerServer2.getDLedgerStore().getLedgerEndIndex());
 
         for (int i = 0; i < 10; i++) {
             GetEntriesResponse getEntriesResponse = dLedgerClient.get(i);
@@ -73,12 +67,9 @@ public class BatchPushTest extends ServerTestHarness{
     public void testBatchPushWithAsyncRequests() throws Exception {
         String group = UUID.randomUUID().toString();
         String peers = String.format("n0-localhost:%d;n1-localhost:%d;n2-localhost:%d", nextPort(), nextPort(), nextPort());
-        DLedgerProxy dLedgerProxy0 = launchDLedgerProxyEnableBatchPush(group, peers, "n0", "n1", DLedgerConfig.FILE);
-        DLedgerProxy dLedgerProxy1 = launchDLedgerProxyEnableBatchPush(group, peers, "n1", "n1", DLedgerConfig.FILE);
-        DLedgerProxy dLedgerProxy2 = launchDLedgerProxyEnableBatchPush(group, peers, "n2", "n1", DLedgerConfig.FILE);
-        DLedgerServer dLedgerServer0 = dLedgerProxy0.getDLedgerManager().getDLedgerServers().get(0);
-        DLedgerServer dLedgerServer1 = dLedgerProxy1.getDLedgerManager().getDLedgerServers().get(0);
-        DLedgerServer dLedgerServer2 = dLedgerProxy2.getDLedgerManager().getDLedgerServers().get(0);
+        DLedgerServer dLedgerServer0 = launchServerEnableBatchPush(group, peers, "n0", "n1", DLedgerConfig.FILE);
+        DLedgerServer dLedgerServer1 = launchServerEnableBatchPush(group, peers, "n1", "n1", DLedgerConfig.FILE);
+        DLedgerServer dLedgerServer2 = launchServerEnableBatchPush(group, peers, "n2", "n1", DLedgerConfig.FILE);
         List<CompletableFuture<AppendEntryResponse>> futures = new ArrayList<>();
         for (int i = 0; i < 10; i++) {
             AppendEntryRequest request = new AppendEntryRequest();
@@ -114,8 +105,7 @@ public class BatchPushTest extends ServerTestHarness{
         String group = UUID.randomUUID().toString();
         String peers = String.format("n0-localhost:%d;n1-localhost:%d", nextPort(), nextPort());
 
-        DLedgerProxy dLedgerProxy0 = launchDLedgerProxyEnableBatchPush(group, peers, "n0", "n0", DLedgerConfig.FILE);
-        DLedgerServer dLedgerServer0 = dLedgerProxy0.getDLedgerManager().getDLedgerServers().get(0);
+        DLedgerServer dLedgerServer0 = launchServerEnableBatchPush(group, peers, "n0", "n0", DLedgerConfig.FILE);
         List<CompletableFuture<AppendEntryResponse>> futures = new ArrayList<>();
         for (int i = 0; i < 10; i++) {
             AppendEntryRequest appendEntryRequest = new AppendEntryRequest();
@@ -156,8 +146,7 @@ public class BatchPushTest extends ServerTestHarness{
         String group = UUID.randomUUID().toString();
         String peers = String.format("n0-localhost:%d;n1-localhost:%d", nextPort(), nextPort());
 
-        DLedgerProxy dLedgerProxy0 = launchDLedgerProxyEnableBatchPush(group, peers, "n0", "n0", DLedgerConfig.FILE);
-        DLedgerServer dLedgerServer0 = dLedgerProxy0.getDLedgerManager().getDLedgerServers().get(0);
+        DLedgerServer dLedgerServer0 = launchServerEnableBatchPush(group, peers, "n0", "n0", DLedgerConfig.FILE);
         AtomicBoolean sendSuccess = new AtomicBoolean(false);
         AppendEntryRequest appendEntryRequest = new AppendEntryRequest();
         appendEntryRequest.setGroup(group);
@@ -171,11 +160,11 @@ public class BatchPushTest extends ServerTestHarness{
         Thread.sleep(500);
         Assertions.assertTrue(!sendSuccess.get());
         //start server1
-        DLedgerProxy dLedgerProxy1 = launchDLedgerProxyEnableBatchPush(group, peers, "n1", "n0", DLedgerConfig.FILE);
+        DLedgerServer dLedgerServer1 = launchServerEnableBatchPush(group, peers, "n1", "n0", DLedgerConfig.FILE);
         Thread.sleep(1500);
         Assertions.assertTrue(sendSuccess.get());
         //shutdown server1
-        dLedgerProxy1.shutdown();
+        dLedgerServer1.shutdown();
         sendSuccess.set(false);
         future = dLedgerServer0.handleAppend(appendEntryRequest);
         Assertions.assertTrue(future instanceof AppendFuture);
@@ -185,8 +174,7 @@ public class BatchPushTest extends ServerTestHarness{
         Thread.sleep(500);
         Assertions.assertTrue(!sendSuccess.get());
         //restart server1
-        dLedgerProxy1 = launchDLedgerProxyEnableBatchPush(group, peers, "n1", "n0", DLedgerConfig.FILE);
-        DLedgerServer dLedgerServer1 = dLedgerProxy1.getDLedgerManager().getDLedgerServers().get(0);
+        dLedgerServer1 = launchServerEnableBatchPush(group, peers, "n1", "n0", DLedgerConfig.FILE);
         Thread.sleep(1500);
         Assertions.assertTrue(sendSuccess.get());
 
@@ -200,20 +188,19 @@ public class BatchPushTest extends ServerTestHarness{
     public void testBatchPushMissed() throws Exception {
         String group = UUID.randomUUID().toString();
         String peers = String.format("n0-localhost:%d;n1-localhost:%d", nextPort(), nextPort());
-        DLedgerProxy dLedgerProxy0 = launchDLedgerProxyEnableBatchPush(group, peers, "n0", "n0", DLedgerConfig.FILE);
-        DLedgerProxy dLedgerProxy1 = launchDLedgerProxyEnableBatchPush(group, peers, "n1", "n0", DLedgerConfig.FILE);
-        DLedgerProxy mockProxy1 = Mockito.spy(dLedgerProxy1);
+        DLedgerServer dLedgerServer0 = launchServerEnableBatchPush(group, peers, "n0", "n0", DLedgerConfig.FILE);
+        DLedgerServer dLedgerServer1 = launchServerEnableBatchPush(group, peers, "n1", "n0", DLedgerConfig.FILE);
+        DLedgerServer mockServer1 = Mockito.spy(dLedgerServer1);
         AtomicInteger callNum = new AtomicInteger(0);
         doAnswer(x -> {
             if (callNum.incrementAndGet() % 3 == 0) {
                 return new CompletableFuture<>();
             } else {
-                return dLedgerProxy1.handlePush(x.getArgument(0));
+                return dLedgerServer1.handlePush(x.getArgument(0));
             }
-        }).when(mockProxy1).handlePush(any());
-        ((DLedgerRpcNettyService) dLedgerProxy1.getdLedgerRpcService()).setdLedgerProxy(mockProxy1);
-        DLedgerServer dLedgerServer0 = dLedgerProxy0.getDLedgerManager().getDLedgerServers().get(0);
-        DLedgerServer dLedgerServer1 = dLedgerProxy1.getDLedgerManager().getDLedgerServers().get(0);
+        }).when(mockServer1).handlePush(any());
+        ((DLedgerRpcNettyService) dLedgerServer1.getDLedgerRpcService()).setDLedger(mockServer1);
+
         for (int i = 0; i < 10; i++) {
             AppendEntryRequest appendEntryRequest = new AppendEntryRequest();
             appendEntryRequest.setGroup(group);
@@ -234,8 +221,7 @@ public class BatchPushTest extends ServerTestHarness{
     public void testBatchPushTruncate() throws Exception {
         String group = UUID.randomUUID().toString();
         String peers = String.format("n0-localhost:%d;n1-localhost:%d", nextPort(), nextPort());
-        DLedgerProxy dLedgerProxy0 = launchDLedgerProxyEnableBatchPush(group, peers, "n0", "n0", DLedgerConfig.FILE);
-        DLedgerServer dLedgerServer0 = dLedgerProxy0.getDLedgerManager().getDLedgerServers().get(0);
+        DLedgerServer dLedgerServer0 = launchServerEnableBatchPush(group, peers, "n0", "n0", DLedgerConfig.FILE);
         for (int i = 0; i < 10; i++) {
             DLedgerEntry entry = new DLedgerEntry();
             entry.setBody(new byte[128]);
@@ -248,21 +234,18 @@ public class BatchPushTest extends ServerTestHarness{
         for (long i = 0; i < 10; i++) {
             entries.add(dLedgerServer0.getDLedgerStore().get(i));
         }
-        dLedgerProxy0.shutdown();
+        dLedgerServer0.shutdown();
 
-        DLedgerProxy dLedgerProxy1 = launchDLedgerProxyEnableBatchPush(group, peers, "n1", "n0", DLedgerConfig.FILE);
-        DLedgerServer dLedgerServer1 = dLedgerProxy1.getDLedgerManager().getDLedgerServers().get(0);
+        DLedgerServer dLedgerServer1 = launchServerEnableBatchPush(group, peers, "n1", "n0", DLedgerConfig.FILE);
         for (int i = 0; i < 5; i++) {
             DLedgerEntry resEntry = dLedgerServer1.getDLedgerStore().appendAsFollower(entries.get(i), 0, "n0");
             Assertions.assertEquals(i, resEntry.getIndex());
         }
-        dLedgerProxy1.shutdown();
+        dLedgerServer1.shutdown();
 
         //change leader from n0 => n1
-        dLedgerProxy1 = launchDLedgerProxyEnableBatchPush(group, peers, "n1", "n1", DLedgerConfig.FILE);
-        dLedgerProxy0 = launchDLedgerProxyEnableBatchPush(group, peers, "n0", "n1", DLedgerConfig.FILE);
-        dLedgerServer0 = dLedgerProxy0.getDLedgerManager().getDLedgerServers().get(0);
-        dLedgerServer1 = dLedgerProxy1.getDLedgerManager().getDLedgerServers().get(0);
+        dLedgerServer1 = launchServerEnableBatchPush(group, peers, "n1", "n1", DLedgerConfig.FILE);
+        dLedgerServer0 = launchServerEnableBatchPush(group, peers, "n0", "n1", DLedgerConfig.FILE);
         Thread.sleep(1000);
         Assertions.assertEquals(0, dLedgerServer0.getDLedgerStore().getLedgerBeginIndex());
         Assertions.assertEquals(4, dLedgerServer0.getDLedgerStore().getLedgerEndIndex());
