+++ dledger/src/test/java/io/openmessaging/storage/dledger/ServerTestHarness.java
@@ -17,12 +17,8 @@
 package io.openmessaging.storage.dledger;
 
 import io.openmessaging.storage.dledger.client.DLedgerClient;
-import io.openmessaging.storage.dledger.dledger.DLedgerProxy;
-import io.openmessaging.storage.dledger.dledger.DLedgerProxyConfig;
 import io.openmessaging.storage.dledger.util.FileTestUtil;
-
 import java.io.File;
-import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
 
@@ -52,7 +48,7 @@ public class ServerTestHarness extends ServerTestBase {
     }
 
     protected synchronized DLedgerServer launchServer(String group, String peers, String selfId, String leaderId,
-                                                      String storeType) {
+        String storeType) {
         DLedgerConfig config = new DLedgerConfig();
         config.group(group).selfId(selfId).peers(peers);
         config.setStoreBaseDir(FileTestUtil.TEST_BASE + File.separator + group);
@@ -74,127 +70,10 @@ public class ServerTestHarness extends ServerTestBase {
         bases.add(config.getDefaultPath());
         dLedgerServer.startup();
         return dLedgerServer;
-    }
-
-    protected synchronized DLedgerProxy launchDLedgerProxy(String group, String peers, String selfId) {
-        DLedgerConfig config = new DLedgerConfig();
-        config.setStoreBaseDir(FileTestUtil.TEST_BASE + File.separator + group);
-        config.group(group).selfId(selfId).peers(peers);
-        config.setStoreType(DLedgerConfig.MEMORY);
-        DLedgerProxy dLedgerProxy = new DLedgerProxy(config);
-        dLedgerProxy.startup();
-        bases.add(config.getDefaultPath());
-        return dLedgerProxy;
-    }
-
-    protected synchronized DLedgerConfig createDLedgerConfig(String group, String peers, String selfId) {
-        DLedgerConfig config = new DLedgerConfig();
-        config.setStoreBaseDir(FileTestUtil.TEST_BASE + File.separator + group);
-        config.group(group).selfId(selfId).peers(peers);
-        config.setStoreType(DLedgerConfig.MEMORY);
-        bases.add(config.getDefaultPath());
-        return config;
-    }
-
-
-    protected synchronized DLedgerProxy launchDLedgerProxy(String group, String peers, String selfId, String preferredLeaderId) {
-        DLedgerConfig config = createDLedgerConfig(group, peers, selfId, preferredLeaderId);
-        DLedgerProxy dLedgerProxy = new DLedgerProxy(config);
-        dLedgerProxy.startup();
-        return dLedgerProxy;
-    }
-
-    protected synchronized DLedgerConfig createDLedgerConfig(String group, String peers, String selfId, String preferredLeaderId) {
-        DLedgerConfig config = new DLedgerConfig();
-        config.setStoreBaseDir(FileTestUtil.TEST_BASE + File.separator + group);
-        config.group(group).selfId(selfId).peers(peers);
-        config.setStoreType(DLedgerConfig.MEMORY);
-        config.setPreferredLeaderIds(preferredLeaderId);
-        bases.add(config.getDefaultPath());
-        return config;
-    }
-
-    protected synchronized DLedgerProxy launchDLedgerProxy(String group, String peers, String selfId, String leaderId,
-                                                           String storeType) {
-        DLedgerConfig config = new DLedgerConfig();
-        config.group(group).selfId(selfId).peers(peers);
-        config.setStoreBaseDir(FileTestUtil.TEST_BASE + File.separator + group);
-        config.setStoreType(storeType);
-        config.setMappedFileSizeForEntryData(10 * 1024 * 1024);
-        config.setEnableLeaderElector(false);
-        config.setEnableDiskForceClean(false);
-        config.setDiskSpaceRatioToForceClean(0.90f);
-        DLedgerProxy dLedgerProxy = new DLedgerProxy(config);
-        MemberState memberState = dLedgerProxy.getDLedgerManager().getDLedgerServers().get(0).getMemberState();
-        memberState.setCurrTermForTest(0);
-        if (selfId.equals(leaderId)) {
-            memberState.changeToLeader(0);
-        } else {
-            memberState.changeToFollower(0, leaderId);
-        }
-        bases.add(config.getDataStorePath());
-        bases.add(config.getIndexStorePath());
-        bases.add(config.getDefaultPath());
-        dLedgerProxy.startup();
-        return dLedgerProxy;
-    }
-
-    protected synchronized DLedgerConfig createDLedgerConfig(String group, String peers, String selfId, String leaderId,
-                                                             String storeType) {
-        DLedgerConfig config = new DLedgerConfig();
-        config.group(group).selfId(selfId).peers(peers);
-        config.setStoreBaseDir(FileTestUtil.TEST_BASE + File.separator + group);
-        config.setStoreType(storeType);
-        config.setMappedFileSizeForEntryData(10 * 1024 * 1024);
-        config.setEnableLeaderElector(false);
-        config.setEnableDiskForceClean(false);
-        config.setDiskSpaceRatioToForceClean(0.90f);
-        config.setPreferredLeaderIds(leaderId);
-        bases.add(config.getDataStorePath());
-        bases.add(config.getIndexStorePath());
-        bases.add(config.getDefaultPath());
-        return config;
-    }
-
-    protected synchronized DLedgerProxy launchDLedgerProxy(DLedgerProxyConfig dLedgerProxyConfig) {
-        DLedgerProxy dLedgerProxy = new DLedgerProxy(dLedgerProxyConfig.getConfigs());
-        for (DLedgerServer dLedgerServer : dLedgerProxy.getDLedgerManager().getDLedgerServers()) {
-            String leaderId = dLedgerServer.getdLedgerConfig().getPreferredLeaderIds();
-            if (!dLedgerServer.getdLedgerConfig().isEnableLeaderElector() && leaderId != null) {
-                dLedgerServer.getMemberState().setCurrTermForTest(0);
-                if (dLedgerServer.getMemberState().getSelfId().equals(leaderId)) {
-                    dLedgerServer.getMemberState().changeToLeader(0);
-                } else {
-                    dLedgerServer.getMemberState().changeToFollower(0, leaderId);
-                }
-            }
-        }
-        dLedgerProxy.startup();
-        return dLedgerProxy;
-    }
-
-    protected synchronized DLedgerProxy[] launchDLedgerProxy(DLedgerProxyConfig[] dLedgerProxyConfigs) {
-        DLedgerProxy[] proxies = new DLedgerProxy[dLedgerProxyConfigs.length];
-        for (int i = 0; i < dLedgerProxyConfigs.length; i++) {
-            DLedgerProxy dLedgerProxy = new DLedgerProxy(dLedgerProxyConfigs[i]);
-            for (DLedgerServer dLedgerServer : dLedgerProxy.getDLedgerManager().getDLedgerServers()) {
-                String leaderId = dLedgerServer.getdLedgerConfig().getPreferredLeaderIds();
-                if (!dLedgerServer.getdLedgerConfig().isEnableLeaderElector() && leaderId != null) {
-                    if (dLedgerServer.getMemberState().getSelfId().equals(leaderId)) {
-                        dLedgerServer.getMemberState().changeToLeader(0);
-                    } else {
-                        dLedgerServer.getMemberState().changeToFollower(0, leaderId);
-                    }
-                }
-            }
-            dLedgerProxy.startup();
-            proxies[i] = dLedgerProxy;
-        }
-        return proxies;
     }
 
     protected synchronized DLedgerServer launchServerEnableBatchPush(String group, String peers, String selfId, String leaderId,
-                                                                     String storeType) {
+        String storeType) {
         DLedgerConfig config = new DLedgerConfig();
         config.group(group).selfId(selfId).peers(peers);
         config.setStoreBaseDir(FileTestUtil.TEST_BASE + File.separator + group);
@@ -218,35 +97,6 @@ public class ServerTestHarness extends ServerTestBase {
         bases.add(config.getDefaultPath());
         dLedgerServer.startup();
         return dLedgerServer;
-    }
-
-    protected synchronized DLedgerProxy launchDLedgerProxyEnableBatchPush(String group, String peers, String selfId, String leaderId,
-                                                                          String storeType) {
-        DLedgerProxyConfig dLedgerProxyConfig = new DLedgerProxyConfig();
-        DLedgerConfig config = new DLedgerConfig();
-        config.group(group).selfId(selfId).peers(peers);
-        config.setStoreBaseDir(FileTestUtil.TEST_BASE + File.separator + group);
-        config.setStoreType(storeType);
-        config.setMappedFileSizeForEntryData(10 * 1024 * 1024);
-        config.setEnableLeaderElector(false);
-        config.setEnableDiskForceClean(false);
-        config.setDiskSpaceRatioToForceClean(0.90f);
-        config.setEnableBatchPush(true);
-        config.setMaxBatchPushSize(300);
-        dLedgerProxyConfig.setConfigs(Arrays.asList(config));
-        DLedgerProxy dLedgerProxy = new DLedgerProxy(dLedgerProxyConfig);
-        MemberState memberState = dLedgerProxy.getDLedgerManager().getDLedgerServer(config.getGroup(), config.getSelfId()).getMemberState();
-        memberState.setCurrTermForTest(0);
-        if (selfId.equals(leaderId)) {
-            memberState.changeToLeader(0);
-        } else {
-            memberState.changeToFollower(0, leaderId);
-        }
-        bases.add(config.getDataStorePath());
-        bases.add(config.getIndexStorePath());
-        bases.add(config.getDefaultPath());
-        dLedgerProxy.startup();
-        return dLedgerProxy;
     }
 
     protected synchronized DLedgerClient launchClient(String group, String peers) {
@@ -256,13 +106,12 @@ public class ServerTestHarness extends ServerTestBase {
     }
 
     protected DLedgerServer parseServers(List<DLedgerServer> servers, AtomicInteger leaderNum,
-                                         AtomicInteger followerNum) {
+        AtomicInteger followerNum) {
         DLedgerServer leaderServer = null;
         for (DLedgerServer server : servers) {
             if (server.getMemberState().isLeader()) {
                 leaderNum.incrementAndGet();
                 leaderServer = server;
-                System.out.println("server:"+server.getdLedgerConfig().getSelfId());
             } else if (server.getMemberState().isFollower()) {
                 followerNum.incrementAndGet();
             }
