+++ dledger/src/test/java/io/openmessaging/storage/dledger/protocol/HeartbeatRequestTest.java
@@ -18,7 +18,6 @@ package io.openmessaging.storage.dledger.protocol;
 
 import io.openmessaging.storage.dledger.DLedgerServer;
 import io.openmessaging.storage.dledger.ServerTestHarness;
-import io.openmessaging.storage.dledger.dledger.DLedgerProxy;
 import io.openmessaging.storage.dledger.utils.DLedgerUtils;
 import java.util.UUID;
 import org.junit.jupiter.api.Assertions;
@@ -30,10 +29,8 @@ public class HeartbeatRequestTest extends ServerTestHarness {
     public void testHeartbeat() throws Exception {
         String group = UUID.randomUUID().toString();
         String peers = String.format("n0-localhost:%d;n1-localhost:%d;n2-localhost:%d", nextPort(), nextPort(), nextPort());
-        DLedgerProxy dLedgerProxy0 = launchDLedgerProxy(group, peers, "n0");
-        DLedgerProxy dLedgerProxy1 = launchDLedgerProxy(group, peers, "n1");
-        DLedgerServer dLedgerServer0 = dLedgerProxy0.getDLedgerManager().getDLedgerServers().get(0);
-        DLedgerServer dLedgerServer1 = dLedgerProxy1.getDLedgerManager().getDLedgerServers().get(0);
+        DLedgerServer dLedgerServer0 = launchServer(group, peers, "n0");
+        DLedgerServer dLedgerServer1 = launchServer(group, peers, "n1");
         DLedgerServer leader, follower;
         {
             long start = System.currentTimeMillis();
@@ -100,8 +97,8 @@ public class HeartbeatRequestTest extends ServerTestHarness {
             Thread.sleep(100);
             Assertions.assertEquals(DLedgerResponseCode.SUCCESS.getCode(), follower.handleHeartBeat(request).get().getCode());
         }
-        dLedgerProxy0.shutdown();
-        dLedgerProxy1.shutdown();
+        dLedgerServer0.shutdown();
+        dLedgerServer1.shutdown();
     }
 
 }
