+++ proxy/src/test/java/io/openmessaging/storage/dledger/AppendAndPushTest.java
@@ -21,6 +21,7 @@ import io.openmessaging.storage.dledger.protocol.AppendEntryRequest;
 import io.openmessaging.storage.dledger.protocol.AppendEntryResponse;
 import io.openmessaging.storage.dledger.protocol.BatchAppendEntryRequest;
 import io.openmessaging.storage.dledger.protocol.DLedgerResponseCode;
+import io.openmessaging.storage.dledger.proxy.DLedgerProxy;
 import io.openmessaging.storage.dledger.utils.DLedgerUtils;
 import java.util.ArrayList;
 import java.util.LinkedList;
@@ -44,7 +45,8 @@ public class AppendAndPushTest extends ServerTestHarness {
         String group = UUID.randomUUID().toString();
         String peers = String.format("n0-localhost:%d;n1-localhost:%d", nextPort(), nextPort());
 
-        DLedgerServer dLedgerServer0 = launchServer(group, peers, "n0", "n0", DLedgerConfig.FILE);
+        DLedgerProxy dLedgerProxy0 = launchDLedgerProxy(group, peers, "n0", "n0", DLedgerConfig.FILE);
+        DLedgerServer dLedgerServer0 = dLedgerProxy0.getDLedgerManager().getDLedgerServers().get(0);
         List<CompletableFuture<AppendEntryResponse>> futures = new ArrayList<>();
         for (int i = 0; i < 10; i++) {
             AppendEntryRequest appendEntryRequest = new AppendEntryRequest();
@@ -55,9 +57,10 @@ public class AppendAndPushTest extends ServerTestHarness {
             Assertions.assertTrue(future instanceof AppendFuture);
             futures.add(future);
         }
-        Assertions.assertEquals(9, dLedgerServer0.getDLedgerStore().getLedgerEndIndex());
-        Assertions.assertEquals(-1, dLedgerServer0.getDLedgerStore().getCommittedIndex());
-        DLedgerServer dLedgerServer1 = launchServer(group, peers, "n1", "n0", DLedgerConfig.FILE);
+        Assertions.assertEquals(9, dLedgerServer0.getdLedgerStore().getLedgerEndIndex());
+        Assertions.assertEquals(-1, dLedgerServer0.getdLedgerStore().getCommittedIndex());
+        DLedgerProxy dLedgerProxy1 = launchDLedgerProxy(group, peers, "n1", "n0", DLedgerConfig.FILE);
+        DLedgerServer dLedgerServer1 = dLedgerProxy1.getDLedgerManager().getDLedgerServers().get(0);
         long start = System.currentTimeMillis();
         while (DLedgerUtils.elapsed(start) < 3000 && dLedgerServer1.getDLedgerStore().getCommittedIndex() != 9) {
             DLedgerUtils.sleep(100);
@@ -70,8 +73,8 @@ public class AppendAndPushTest extends ServerTestHarness {
     public void testPushNetworkOffline() throws Exception {
         String group = UUID.randomUUID().toString();
         String peers = String.format("n0-localhost:%d;n1-localhost:%d", nextPort(), nextPort());
-
-        DLedgerServer dLedgerServer0 = launchServer(group, peers, "n0", "n0", DLedgerConfig.FILE);
+        DLedgerProxy dLedgerProxy0 = launchDLedgerProxy(group, peers, "n0", "n0", DLedgerConfig.FILE);
+        DLedgerServer dLedgerServer0 = dLedgerProxy0.getDLedgerManager().getDLedgerServers().get(0);
         List<CompletableFuture<AppendEntryResponse>> futures = new ArrayList<>();
         for (int i = 0; i < 10; i++) {
             AppendEntryRequest appendEntryRequest = new AppendEntryRequest();
@@ -112,7 +115,8 @@ public class AppendAndPushTest extends ServerTestHarness {
         String group = UUID.randomUUID().toString();
         String peers = String.format("n0-localhost:%d;n1-localhost:%d", nextPort(), nextPort());
 
-        DLedgerServer dLedgerServer0 = launchServer(group, peers, "n0", "n0", DLedgerConfig.FILE);
+        DLedgerProxy dLedgerProxy0 = launchDLedgerProxy(group, peers, "n0", "n0", DLedgerConfig.FILE);
+        DLedgerServer dLedgerServer0 = dLedgerProxy0.getDLedgerManager().getDLedgerServers().get(0);
         AtomicBoolean sendSuccess = new AtomicBoolean(false);
         AppendEntryRequest appendEntryRequest = new AppendEntryRequest();
         appendEntryRequest.setGroup(group);
@@ -126,11 +130,11 @@ public class AppendAndPushTest extends ServerTestHarness {
         Thread.sleep(500);
         Assertions.assertTrue(!sendSuccess.get());
         //start server1
-        DLedgerServer dLedgerServer1 = launchServer(group, peers, "n1", "n0", DLedgerConfig.FILE);
+        DLedgerProxy dLedgerProxy1 = launchDLedgerProxy(group, peers, "n1", "n0", DLedgerConfig.FILE);
         Thread.sleep(1500);
         Assertions.assertTrue(sendSuccess.get());
         //shutdown server1
-        dLedgerServer1.shutdown();
+        dLedgerProxy1.shutdown();
         sendSuccess.set(false);
         future = dLedgerServer0.handleAppend(appendEntryRequest);
         Assertions.assertTrue(future instanceof AppendFuture);
@@ -140,7 +144,8 @@ public class AppendAndPushTest extends ServerTestHarness {
         Thread.sleep(500);
         Assertions.assertTrue(!sendSuccess.get());
         //restart server1
-        dLedgerServer1 = launchServer(group, peers, "n1", "n0", DLedgerConfig.FILE);
+        dLedgerProxy1 = launchDLedgerProxy(group, peers, "n1", "n0", DLedgerConfig.FILE);
+        DLedgerServer dLedgerServer1 = dLedgerProxy1.getDLedgerManager().getDLedgerServers().get(0);
         Thread.sleep(1500);
         Assertions.assertTrue(sendSuccess.get());
 
@@ -154,19 +159,20 @@ public class AppendAndPushTest extends ServerTestHarness {
     public void testPushMissed() throws Exception {
         String group = UUID.randomUUID().toString();
         String peers = String.format("n0-localhost:%d;n1-localhost:%d", nextPort(), nextPort());
-        DLedgerServer dLedgerServer0 = launchServer(group, peers, "n0", "n0", DLedgerConfig.FILE);
-        DLedgerServer dLedgerServer1 = launchServer(group, peers, "n1", "n0", DLedgerConfig.FILE);
-        DLedgerServer mockServer1 = Mockito.spy(dLedgerServer1);
+        DLedgerProxy dLedgerProxy0 = launchDLedgerProxy(group, peers, "n0", "n0", DLedgerConfig.FILE);
+        DLedgerProxy dLedgerProxy1 = launchDLedgerProxy(group, peers, "n1", "n0", DLedgerConfig.FILE);
+        DLedgerProxy mockProxy1 = Mockito.spy(dLedgerProxy1);
         AtomicInteger callNum = new AtomicInteger(0);
         doAnswer(x -> {
             if (callNum.incrementAndGet() % 3 == 0) {
                 return new CompletableFuture<>();
             } else {
-                return dLedgerServer1.handlePush(x.getArgument(0));
+                return dLedgerProxy1.handlePush(x.getArgument(0));
             }
-        }).when(mockServer1).handlePush(any());
-        ((DLedgerRpcNettyService) dLedgerServer1.getDLedgerRpcService()).setDLedger(mockServer1);
-
+        }).when(mockProxy1).handlePush(any());
+        ((DLedgerRpcNettyService) dLedgerProxy1.getDLedgerRpcService()).setDLedger(mockProxy1);
+        DLedgerServer dLedgerServer0 = dLedgerProxy0.getDLedgerManager().getDLedgerServers().get(0);
+        DLedgerServer dLedgerServer1 = dLedgerProxy1.getDLedgerManager().getDLedgerServers().get(0);
         for (int i = 0; i < 10; i++) {
             AppendEntryRequest appendEntryRequest = new AppendEntryRequest();
             appendEntryRequest.setGroup(group);
@@ -187,7 +193,8 @@ public class AppendAndPushTest extends ServerTestHarness {
     public void testPushTruncate() throws Exception {
         String group = UUID.randomUUID().toString();
         String peers = String.format("n0-localhost:%d;n1-localhost:%d", nextPort(), nextPort());
-        DLedgerServer dLedgerServer0 = launchServer(group, peers, "n0", "n0", DLedgerConfig.FILE);
+        DLedgerProxy dLedgerProxy0 = launchDLedgerProxy(group, peers, "n0", "n0", DLedgerConfig.FILE);
+        DLedgerServer dLedgerServer0 = dLedgerProxy0.getDLedgerManager().getDLedgerServers().get(0);
         for (int i = 0; i < 10; i++) {
             DLedgerEntry entry = new DLedgerEntry();
             entry.setBody(new byte[128]);
@@ -200,18 +207,20 @@ public class AppendAndPushTest extends ServerTestHarness {
         for (long i = 0; i < 10; i++) {
             entries.add(dLedgerServer0.getDLedgerStore().get(i));
         }
-        dLedgerServer0.shutdown();
+        dLedgerProxy0.shutdown();
 
-        DLedgerServer dLedgerServer1 = launchServer(group, peers, "n1", "n0", DLedgerConfig.FILE);
+        DLedgerProxy dLedgerProxy1 = launchDLedgerProxy(group, peers, "n1", "n0", DLedgerConfig.FILE);
+        DLedgerServer dLedgerServer1 = dLedgerProxy1.getDLedgerManager().getDLedgerServers().get(0);
         for (int i = 0; i < 5; i++) {
             DLedgerEntry resEntry = dLedgerServer1.getDLedgerStore().appendAsFollower(entries.get(i), 0, "n0");
             Assertions.assertEquals(i, resEntry.getIndex());
         }
-        dLedgerServer1.shutdown();
-
+        dLedgerProxy1.shutdown();
         //change leader from n0 => n1
-        dLedgerServer1 = launchServer(group, peers, "n1", "n1", DLedgerConfig.FILE);
-        dLedgerServer0 = launchServer(group, peers, "n0", "n1", DLedgerConfig.FILE);
+        dLedgerProxy1 = launchDLedgerProxy(group, peers, "n1", "n1", DLedgerConfig.FILE);
+        dLedgerProxy0 = launchDLedgerProxy(group, peers, "n0", "n1", DLedgerConfig.FILE);
+        dLedgerServer0 = dLedgerProxy0.getDLedgerManager().getDLedgerServers().get(0);
+        dLedgerServer1 = dLedgerProxy1.getDLedgerManager().getDLedgerServers().get(0);
         Thread.sleep(1000);
         Assertions.assertEquals(0, dLedgerServer0.getDLedgerStore().getLedgerBeginIndex());
         Assertions.assertEquals(4, dLedgerServer0.getDLedgerStore().getLedgerEndIndex());
@@ -231,13 +240,13 @@ public class AppendAndPushTest extends ServerTestHarness {
     public void testBatchAppend() throws Exception {
         String group = UUID.randomUUID().toString();
         String peers = String.format("n0-localhost:%d;n1-localhost:%d", nextPort(), nextPort());
-        DLedgerServer dLedgerServer0 = launchServer(group, peers, "n0", "n0", DLedgerConfig.FILE);
-        DLedgerServer dLedgerServer1 = launchServer(group, peers, "n1", "n0", DLedgerConfig.FILE);
-        DLedgerServer mockServer1 = Mockito.spy(dLedgerServer1);
-        doAnswer(x -> dLedgerServer1.handlePush(x.getArgument(0))).when(mockServer1).handlePush(any());
-        ((DLedgerRpcNettyService) dLedgerServer1.getDLedgerRpcService()).setDLedger(mockServer1);
-
-
+        DLedgerProxy dLedgerProxy0 = launchDLedgerProxy(group, peers, "n0", "n0", DLedgerConfig.FILE);
+        DLedgerProxy dLedgerProxy1 = launchDLedgerProxy(group, peers, "n1", "n0", DLedgerConfig.FILE);
+        DLedgerServer dLedgerServer0 = dLedgerProxy0.getDLedgerManager().getDLedgerServers().get(0);
+        DLedgerServer dLedgerServer1 = dLedgerProxy1.getDLedgerManager().getDLedgerServers().get(0);
+        DLedgerProxy mockProxy1 = Mockito.spy(dLedgerProxy1);
+        doAnswer(x -> dLedgerProxy1.handlePush(x.getArgument(0))).when(mockProxy1).handlePush(any());
+        ((DLedgerRpcNettyService) dLedgerServer1.getdLedgerRpcService()).setDLedger(mockProxy1);
         BatchAppendEntryRequest appendEntryRequest = new BatchAppendEntryRequest();
         appendEntryRequest.setGroup(group);
         appendEntryRequest.setRemoteId(dLedgerServer0.getMemberState().getSelfId());
