+++ proxy/src/test/java/io/openmessaging/storage/dledger/util/ConfigUtilsTest.java
@@ -1,8 +1,23 @@
+/*
+ * Copyright 2017-2022 The DLedger Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package io.openmessaging.storage.dledger.util;
 
-import io.openmessaging.storage.dledger.dledger.DLedgerProxyConfig;
-import io.openmessaging.storage.dledger.utils.ConfigUtils;
-import java.io.File;
+import io.openmessaging.storage.dledger.proxy.DLedgerProxyConfig;
+import io.openmessaging.storage.dledger.proxy.util.ConfigUtils;
 import java.nio.file.NoSuchFileException;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
